1st attempt at disabling deletion
Let's see how far this gets me

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@100432 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ extensions/Contest/includes/Contest.class.php
@@ -412,6 +412,11 @@ class Contest extends ContestDBObject {
 	 * @return boolean Success indicator
 	 */
 	public function removeAllFromDB() {
+		global $wgContestDeletionEnabled;
+		if ( !$wgContestDeletionEnabled ) {
+			// Shouldn't get here (UI should prevent it)
+			throw new MWException( 'Contest deletion is disabled', 'contestdeletiondisabled' );
+		}
 		$condition = array( 'contest_id' => $this->getId() );
 
 		$success = ContestChallenge::s()->delete( $condition );
